Move MSSQL tests to jdbc-test module
