--- modules/mssqlserver/src/test/resources/container-license-acceptance.txt
@@ -1,1 +1,0 @@
-microsoft/mssql-server-linux:2017-CU6
